[SKIN] [DIAG]#81 Remove skin error codes definition from motionControlLib/. Create an appropriate skinLib/ directory in lib/icubmod/. Move definitions and skin configuration reader in the newly created directory. Update all relevant include statements.
