+++ src/libraries/icubmod/canBusSkin/CMakeLists.txt
@@ -9,10 +9,10 @@ if(ICUB_HAS_icub_firmware_shared)
 
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}
                        ${YARP_INCLUDE_DIRS} 
-                       ../motionControlLib/
+                       ../skinLib/
                        ${icub_firmware_shared_canProtocolLib_INCLUDE_DIR})
 
-yarp_add_plugin(canBusSkin CanBusSkin.h CanBusSkin.cpp ../motionControlLib/SkinConfigReader.cpp ../motionControlLib/skinDiagnostic.h)
+yarp_add_plugin(canBusSkin CanBusSkin.h CanBusSkin.cpp ../skinLib/SkinConfigReader.cpp ../skinLib/SkinDiagnostics.h)
 target_link_libraries(canBusSkin ${YARP_LIBRARIES} ${ICUB_LIBRARIES})
 
 icub_export_library(canBusSkin)
