+++ src/libraries/icubmod/embObjSkin/CMakeLists.txt
@@ -10,11 +10,11 @@ if (NOT SKIP_embObjSkin)
 
     include_directories(${CMAKE_CURRENT_SOURCE_DIR}
                    ${YARP_INCLUDE_DIRS}
-                   ../motionControlLib
+                   ../skinLib
                    ${iCubDev_INCLUDE_DIRS}
                    ${icub_firmware_shared_canProtocolLib_INCLUDE_DIR})
 
-    yarp_add_plugin(embObjSkin embObjSkin.h embObjSkin.cpp ../motionControlLib/SkinConfigReader.cpp ../motionControlLib/skinDiagnostic.h)
+    yarp_add_plugin(embObjSkin embObjSkin.h embObjSkin.cpp ../skinLib/SkinConfigReader.cpp ../skinLib/SkinDiagnostics.h)
     target_link_libraries(embObjSkin ethResources ${YARP_LIBRARIES})
     icub_export_library(embObjSkin)
 endif (NOT SKIP_embObjSkin)
