+++ src/modules/skinManager/CMakeLists.txt
@@ -16,7 +16,7 @@ SOURCE_GROUP("Header Files" FILES ${folder_header})
 include_directories(${PROJECT_SOURCE_DIR}/include)
 include_directories(${YARP_INCLUDE_DIRS})
 include_directories(${skinDynLib_INCLUDE_DIRS})
-include_directories(${CMAKE_SOURCE_DIR}/src/libraries/icubmod/motionControlLib)
+include_directories(${CMAKE_SOURCE_DIR}/src/libraries/icubmod/skinLib)
 
 
 ADD_EXECUTABLE(${PROJECTNAME} ${folder_source} ${folder_header})
